requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-magic==0.4.27
onnxruntime==1.17.1
numpy>=1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
psutil==5.9.8
requests==2.31.0
